+++ resources/constants.py
@@ -134,6 +134,7 @@ class Constants:
         self.booted_oc_disk:             str = None  #  Determine current disk OCLP booted from
         self.unpack_thread                   = None  #  Determine if unpack thread finished (threading.Thread)
         self.update_stage:               int = 0  #     Determine update stage (see gui_support.py)
+        self.log_filepath:              Path = None  #  Path to log file
 
         self.commit_info: tuple = (None, None, None)  # Commit info (Branch, Commit Date, Commit URL)